refactoring postgeres
